refactor: Send user slowmodes as websocket connects instead of trying to send it in ready payload
